requirements.txt
# Core
streamlit>=1.31.0
numpy>=1.24.0
pillow>=10.0.0
ollama>=0.1.0

# Web automation (Coursera Agent)
playwright>=1.40.0

# AI / Embeddings (Study System)
sentence-transformers>=2.2.0

# Google Docs integration
google-auth>=2.17.0
google-api-python-client>=2.100.0

# OCR — macOS native (best accuracy)
pyobjc-framework-Vision>=10.0; sys_platform == 'darwin'
pyobjc-framework-Quartz>=10.0; sys_platform == 'darwin'

# OCR fallback (cross-platform)
easyocr>=1.7.0

# Image processing
opencv-python>=4.8.0

# Optional but recommended
watchdog>=3.0.0

# Environment / config
python-dotenv>=1.0.0

# UI v2 — FastAPI backend
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
websockets>=12.0
sse-starlette>=2.1.0